+++ flink-python/src/main/java/org/apache/flink/table/runtime/operators/python/aggregate/AbstractPythonStreamAggregateOperator.java
@@ -36,7 +36,7 @@ import org.apache.flink.table.connector.Projection;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.functions.python.PythonAggregateFunctionInfo;
 import org.apache.flink.table.functions.python.PythonEnv;
-import org.apache.flink.table.planner.typeutils.DataViewUtils;
+import org.apache.flink.table.runtime.dataview.DataViewSpec;
 import org.apache.flink.table.runtime.operators.python.AbstractOneInputPythonFunctionOperator;
 import org.apache.flink.table.runtime.operators.python.utils.StreamRecordRowDataWrappingCollector;
 import org.apache.flink.table.runtime.runners.python.beam.BeamTablePythonFunctionRunner;
@@ -66,7 +66,7 @@ public abstract class AbstractPythonStreamAggregateOperator
 
     private final PythonAggregateFunctionInfo[] aggregateFunctions;
 
-    private final DataViewUtils.DataViewSpec[][] dataViewSpecs;
+    private final DataViewSpec[][] dataViewSpecs;
 
     /** The input logical type. */
     protected final RowType inputType;
@@ -125,7 +125,7 @@ public abstract class AbstractPythonStreamAggregateOperator
             RowType inputType,
             RowType outputType,
             PythonAggregateFunctionInfo[] aggregateFunctions,
-            DataViewUtils.DataViewSpec[][] dataViewSpecs,
+            DataViewSpec[][] dataViewSpecs,
             int[] grouping,
             int indexOfCountStar,
             boolean generateUpdateBefore) {
@@ -245,7 +245,7 @@ public abstract class AbstractPythonStreamAggregateOperator
         builder.setMapStateReadCacheSize(mapStateReadCacheSize);
         builder.setMapStateWriteCacheSize(mapStateWriteCacheSize);
         for (int i = 0; i < aggregateFunctions.length; i++) {
-            DataViewUtils.DataViewSpec[] specs = null;
+            DataViewSpec[] specs = null;
             if (i < dataViewSpecs.length) {
                 specs = dataViewSpecs[i];
             }
