+++ flink-python/src/test/java/org/apache/flink/table/runtime/operators/python/aggregate/PythonStreamGroupAggregateOperatorTest.java
@@ -31,7 +31,7 @@ import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.data.StringData;
 import org.apache.flink.table.functions.python.PythonAggregateFunctionInfo;
-import org.apache.flink.table.planner.typeutils.DataViewUtils;
+import org.apache.flink.table.runtime.dataview.DataViewSpec;
 import org.apache.flink.table.runtime.operators.python.scalar.PythonScalarFunctionOperatorTestBase;
 import org.apache.flink.table.runtime.utils.PassThroughStreamAggregatePythonFunctionRunner;
 import org.apache.flink.table.runtime.utils.PythonTestUtils;
@@ -244,7 +244,7 @@ public class PythonStreamGroupAggregateOperatorTest
                     inputType,
                     outputType,
                     aggregateFunctions,
-                    new DataViewUtils.DataViewSpec[0][0],
+                    new DataViewSpec[0][0],
                     grouping,
                     indexOfCountStar,
                     countStarInserted,
