+++ flink-table/flink-table-runtime/src/main/java/org/apache/flink/table/runtime/dataview/DataViewSpec.java
@@ -1,0 +1,56 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.runtime.dataview;
+
+import org.apache.flink.annotation.Internal;
+import org.apache.flink.table.api.dataview.DataView;
+import org.apache.flink.table.types.DataType;
+
+import java.io.Serializable;
+
+/** Information about a {@link DataView} stored in state. */
+@Internal
+public abstract class DataViewSpec implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    private final String stateId;
+
+    private final int fieldIndex;
+
+    private final DataType dataType;
+
+    DataViewSpec(String stateId, int fieldIndex, DataType dataType) {
+        this.stateId = stateId;
+        this.fieldIndex = fieldIndex;
+        this.dataType = dataType;
+    }
+
+    public String getStateId() {
+        return stateId;
+    }
+
+    public int getFieldIndex() {
+        return fieldIndex;
+    }
+
+    public DataType getDataType() {
+        return dataType;
+    }
+}
